requirements.txt
tensorflow==2.15.0
matplotlib
pandas
scipy
torch
torchvision
random
tkinter
pillow